+++ Anchor.toml
@@ -10,3 +10,4 @@ url = "https://anchor.projectserum.com"
 
 [scripts]
 test = "yarn ts-mocha -p ./tsconfig.json -t 100000 tests/test_*.ts"
+
