+++ programs/controller/src/instructions/mod.rs
@@ -1,0 +1,9 @@
+pub mod initialize;
+pub mod mint_uxd;
+pub mod redeem_uxd;
+pub mod register_depository;
+
+pub use initialize::*;
+pub use mint_uxd::*;
+pub use redeem_uxd::*;
+pub use register_depository::*;
